test: Refactor flood tests into a separate test file (#13556)
