+++ test/integration/http2_integration_test.h
@@ -5,14 +5,10 @@
 #include "envoy/config/bootstrap/v3/bootstrap.pb.h"
 #include "envoy/extensions/filters/network/http_connection_manager/v3/http_connection_manager.pb.h"
 
-#include "test/common/http/http2/http2_frame.h"
-#include "test/integration/filters/test_socket_interface.h"
 #include "test/integration/http_integration.h"
 
 #include "absl/synchronization/mutex.h"
 #include "gtest/gtest.h"
-
-using Envoy::Http::Http2::Http2Frame;
 
 namespace Envoy {
 class Http2IntegrationTest : public testing::TestWithParam<Network::Address::IpVersion>,
@@ -71,74 +67,4 @@ public:
   void runHeaderOnlyTest(bool send_request_body, size_t body_size);
 };
 
-class Http2FrameIntegrationTest : public testing::TestWithParam<Network::Address::IpVersion>,
-                                  public HttpIntegrationTest {
-public:
-  Http2FrameIntegrationTest() : HttpIntegrationTest(Http::CodecClient::Type::HTTP2, GetParam()) {}
-
-protected:
-  void startHttp2Session();
-  Http2Frame readFrame();
-  void sendFrame(const Http2Frame& frame);
-  virtual void beginSession();
-
-  IntegrationTcpClientPtr tcp_client_;
-};
-
-class SocketInterfaceSwap {
-public:
-  // Object of this class hold the state determining the IoHandle which
-  // should return EAGAIN from the `writev` call.
-  struct IoHandleMatcher {
-    bool shouldReturnEgain(uint32_t port) const {
-      absl::ReaderMutexLock lock(&mutex_);
-      return port == port_ && writev_returns_egain_;
-    }
-
-    void setSourcePort(uint32_t port) {
-      absl::WriterMutexLock lock(&mutex_);
-      port_ = port;
-    }
-
-    void setWritevReturnsEgain() {
-      absl::WriterMutexLock lock(&mutex_);
-      writev_returns_egain_ = true;
-    }
-
-  private:
-    mutable absl::Mutex mutex_;
-    uint32_t port_ ABSL_GUARDED_BY(mutex_) = 0;
-    bool writev_returns_egain_ ABSL_GUARDED_BY(mutex_) = false;
-  };
-
-  SocketInterfaceSwap();
-  ~SocketInterfaceSwap();
-
-protected:
-  Envoy::Network::SocketInterface* const previous_socket_interface_{
-      Envoy::Network::SocketInterfaceSingleton::getExisting()};
-  std::shared_ptr<IoHandleMatcher> writev_matcher_{std::make_shared<IoHandleMatcher>()};
-  std::unique_ptr<Envoy::Network::SocketInterfaceLoader> test_socket_interface_loader_;
-};
-
-// It is important that the new socket interface is installed before any I/O activity starts and
-// the previous one is restored after all I/O activity stops. Since the HttpIntegrationTest
-// destructor stops Envoy the SocketInterfaceSwap destructor needs to run after it. This order of
-// multiple inheritance ensures that SocketInterfaceSwap destructor runs after
-// Http2FrameIntegrationTest destructor completes.
-class Http2FloodMitigationTest : public SocketInterfaceSwap, public Http2FrameIntegrationTest {
-public:
-  Http2FloodMitigationTest();
-
-protected:
-  void floodServer(const Http2Frame& frame, const std::string& flood_stat, uint32_t num_frames);
-  void floodServer(absl::string_view host, absl::string_view path,
-                   Http2Frame::ResponseStatus expected_http_status, const std::string& flood_stat,
-                   uint32_t num_frames);
-
-  void setNetworkConnectionBufferSize();
-  void beginSession() override;
-  void prefillOutboundDownstreamQueue(uint32_t data_frame_count, uint32_t data_frame_size = 10);
-  void triggerListenerDrain();
-};
 } // namespace Envoy
